exp/requirements.txt
matplotlib==3.6.1
numpy==1.23.4
pillow==9.2.0
pytest==7.1.3
scikit-learn==1.1.2
tqdm>=4.64.1